move more middleware and handlers
