+++ packages/server/lib/pusher.js
@@ -1,8 +1,15 @@
 const Pusher = require('pusher');
 
-module.exports = new Pusher({
+const pusher = new Pusher({
   appId: process.env.PUSHER_APP_ID,
   key: process.env.PUSHER_APP_KEY,
   secret: process.env.PUSHER_APP_SECRET,
   cluster: process.env.PUSHER_CLUSTER,
 });
+
+module.exports = (req, res) => {
+  const socketId = req.body.socket_id;
+  const channel = req.body.channel_name;
+  const auth = pusher.authenticate(socketId, channel);
+  res.send(auth);
+};
